requirements.txt
Flask>=2.0
python-dotenv>=0.19
selenium>=4.8
webdriver-manager>=4.0.0
requests>=2.28
gunicorn>=20.1